+++ lib/plugins/index.js
@@ -13,7 +13,7 @@ module.exports = [
   require('./metrics.js'),
   require('./print.js'),
   require('./remove.js'),
-  require('./rollback/index.js'),
+  require('./rollback.js'),
   require('./slstats/slstats.js'),
   require('./plugin/plugin.js'),
   require('./plugin/install.js'),
