+++ lib/plugins/rollback.js
