+++ test/unit/lib/plugins/rollback.test.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const expect = require('chai').expect;
-const Rollback = require('../../../../../lib/plugins/rollback/index');
-const Serverless = require('../../../../../lib/Serverless');
+const Rollback = require('../../../../lib/plugins/rollback');
+const Serverless = require('../../../../lib/Serverless');
 
 describe('Rollback', () => {
   let rollback;
